day18_LLM_Gateway/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
faiss-cpu
numpy
prometheus-client